refactor(profile): extract handler; route becomes 15-line shim
Profile route had two near-duplicate response builders (cached vs
fresh fetch). Both now flow through buildProfileResponse(), with
getRecentActivity() consolidating the try/catch around event fetch.
The whole flow lives in src/lib/github/profile-handler.ts.

Route /api/profile becomes a re-export. From 271 LOC -> 15 LOC.

Also extends the api-auth-inventory test to follow handler imports:
when a route delegates to @/lib/.../handle*Request, the boundary
marker can live in the handler module. Keeps thin route shims valid
without weakening the inventory invariant.
